sdk-remote: update.
	* sdk-remote: update.
	* src/kernel/connection.hh, src/kernel/connection.cc,
	* src/object/socket.hh, src/object/socket.cc:
	Adjust the signatures of onRead.

+++ src/kernel/connection.hh
@@ -26,7 +26,7 @@ namespace kernel
     enum { PACKET_SIZE = 1000000000 };
 
     // libport::Socket virtual methods.
-    virtual int onRead(const void* data, size_t length);
+    virtual size_t onRead(const void* data, size_t length);
 
     virtual void onConnect();
